backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# AI / LLM
openai==1.59.9
tiktoken==0.8.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.5
pydantic-settings==2.7.1

# HTTP & Webhooks
httpx==0.28.1

# Email
sendgrid==6.11.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
slowapi==0.1.9
